Fixed build errors, ported to qt 5.2.1
