+++ src/plugins/html/.gitignore
@@ -1,0 +1,1 @@
+*.pro.user
