+++ src/plugins/html/html.qrc
