+++ src/plugins/html/qhtmlbackingstore.cpp
