+++ src/plugins/html/qhtmlcursor.cpp
