--- src/plugins/platforms/platforms.pro
@@ -1,3 +1,0 @@
-TEMPLATE = subdirs
-
-SUBDIRS += html
