--- src/services/html/html.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<SFW version="1.1">
-<service>
-	<name>HtmlService</name>
-	<ipcaddress>html_service</ipcaddress>
-	<description>Html Platform Integration Service</description>
-	<interface>
-		<name>com.nokia.qt.qpa.HtmlService</name>
-		<version>1.0</version>
-		<description>Default implementation of HtmlService</description>
-	</interface>
-</service>
-</SFW>
